Remove the EBF_PlugAndPlayDevice class, Change EBF_HalInstance id field to 32bit and use it as PnP ID as well

+++ src/Core/EBF_HalInstance.cpp
@@ -16,7 +16,7 @@ EBF_HalInstance::EBF_HalInstance() {
 	this->numberOfInstances++;
 }
 
-uint8_t EBF_HalInstance::Init(HAL_Type type, uint8_t id)
+uint8_t EBF_HalInstance::Init(HAL_Type type, uint32_t id)
 {
 	this->pollIntervalMs = 0;
 	this->type = type;
